Added navigation bar, making use of a base_page
All background images are now rendered through the base page, and the image address can be provided trough the widget properties. Modified the swipe.png image to fit better with the 'physical' design. Modified all other pages to make use of the new nav-bar, and removed the deprecated shop page.